move nodefilter functionality under Downloader
