+++ tests/test_download.py
@@ -370,13 +370,13 @@ def test_get_stream(api, tmpdir, smallest_online_products):
 
 @pytest.mark.vcr
 @pytest.mark.scihub
-def test_download_product_nodes(products_api, tmpdir, node_test_products):
+def test_download_product_nodes(api, tmpdir, node_test_products):
     uuid = node_test_products[0]["id"]
     product_dir = node_test_products[0]["title"] + ".SAFE"
     expected_path = tmpdir.join(product_dir)
 
     nodefilter = make_path_filter("*preview/*.kml")
-    product_info = products_api.download(uuid, str(tmpdir), checksum=True, nodefilter=nodefilter)
+    product_info = api.download(uuid, str(tmpdir), checksum=True, nodefilter=nodefilter)
 
     assert os.path.normpath(product_info["node_path"]) == product_dir
     assert expected_path.exists()
